src: move VRAM code to VBlankInterruptHandler
The main loop is now more standard:
- When the LCD is displaying the screen, the commands for loading the required data are prepared;
- Wait for VBlank
- The VBlank interrupt handler executes the requested copy commands
- The main loop restarts again

This is closer to how a proper game would be programmed. Also it allows us to better optimise the time spend computing the data to load vs. the time to copy the actual data.

As a bonus, we can now copy a full 128 tiles tileset in one pass, even on SameBoy.